+++ bolts/data/datamodules/__init__.py
@@ -1,0 +1,1 @@
+from .vision import *
