+++ bolts/data/datasets/vision/__init__.py
@@ -1,0 +1,1 @@
+from .mnist import *
