+++ bolts/fair/data/datamodules/__init__.py
@@ -1,0 +1,5 @@
+"""Make accessible."""
+
+from .base import *
+from .tabular import *
+from .vision import *
