+++ bolts/fair/data/datamodules/vision/__init__.py
@@ -1,0 +1,2 @@
+from .celeba import *
+from .cmnist import *
