+++ bolts/fair/data/datasets/__init__.py
@@ -1,0 +1,4 @@
+from .dummy import *
+from .ethicml import *
+from .vision import *
+from .wrappers import *
