--- bolts/fairness/datamodules/__init__.py
@@ -1,6 +1,0 @@
-"""Make accessible."""
-
-from .base import *
-from .tabular import *
-from .vision import *
-from .wrappers import *
