--- bolts/fairness/datasets/__init__.py
@@ -1,3 +1,0 @@
-from .dummy import *
-from .ethicml import *
-from .vision import *
