+++ tests/data_test.py
@@ -1,25 +1,25 @@
-from typing import Type
+# from typing import Type
 
-import pytest
-from pytorch_lightning import LightningDataModule
-import torch
+# import pytest
+# from pytorch_lightning import LightningDataModule
+# import torch
 
-from bolts.datamodules.mnist_datamodule import MNISTDataModule
-
-
-def _create_dm(dm_cls: Type[LightningDataModule]) -> LightningDataModule:
-    dm = dm_cls(batch_size=2)
-    dm.prepare_data()
-    dm.setup()
-    return dm
+# from bolts.datamodules.mnist_datamodule import MNISTDataModule
 
 
-@pytest.mark.slow
-@pytest.mark.parametrize("dm_cls", [MNISTDataModule])
-def test_data_modules(dm_cls: Type[LightningDataModule]) -> None:
-    """Test the datamodules."""
-    dm = _create_dm(dm_cls)
-    loader = dm.train_dataloader()
-    img, _ = next(iter(loader))
-    assert img.size() == torch.Size([2, *dm.size()])
-    assert dm.num_classes
+# def _create_dm(dm_cls: Type[LightningDataModule]) -> LightningDataModule:
+#     dm = dm_cls(batch_size=2)
+#     dm.prepare_data()
+#     dm.setup()
+#     return dm
+
+
+# @pytest.mark.slow
+# @pytest.mark.parametrize("dm_cls", [MNISTDataModule])
+# def test_data_modules(dm_cls: Type[LightningDataModule]) -> None:
+#     """Test the datamodules."""
+#     dm = _create_dm(dm_cls)
+#     loader = dm.train_dataloader()
+#     img, _ = next(iter(loader))
+#     assert img.size() == torch.Size([2, *dm.size()])
+#     assert dm.num_classes
