--- tests/metrics/classification/accuracy_per_sens_test.py
@@ -1,103 +1,0 @@
-"""Accuracy tests."""
-from typing import NamedTuple
-
-import pytest
-import torch
-from torch import Tensor
-
-from bolts.metrics.accuracy_per_sens import AccuracyPerSens
-
-NUM_PROCESSES = 2
-NUM_BATCHES = 10
-BATCH_SIZE = 32
-NUM_CLASSES = 5
-EXTRA_DIM = 3
-THRESHOLD = 0.5
-
-
-class Input(NamedTuple):
-    preds: Tensor
-    sens: Tensor
-    target: Tensor
-
-
-_input_binary_prob = Input(
-    preds=torch.rand(NUM_BATCHES, BATCH_SIZE),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE)),
-)
-
-_input_binary = Input(
-    preds=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE)),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE)),
-)
-
-_input_multilabel_prob = Input(
-    preds=torch.rand(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)),
-)
-
-_input_multilabel_multidim_prob = Input(
-    preds=torch.rand(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)),
-)
-
-_input_multilabel = Input(
-    preds=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)),
-)
-
-_input_multilabel_multidim = Input(
-    preds=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)),
-    sens=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)),
-    target=torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)),
-)
-
-# Generate edge multilabel edge case, where nothing matches (scores are undefined)
-__temp_preds = torch.randint(high=2, size=(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES))
-__temp_target = abs(__temp_preds - 1)
-
-_input_multilabel_no_match = Input(preds=__temp_preds, sens=__temp_target, target=__temp_target)
-
-__mc_prob_preds = torch.rand(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES)
-__mc_prob_preds = __mc_prob_preds / __mc_prob_preds.sum(dim=2, keepdim=True)
-
-_input_multiclass_prob = Input(
-    preds=__mc_prob_preds,
-    sens=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE)),
-    target=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE)),
-)
-
-_input_multiclass = Input(
-    preds=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE)),
-    sens=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE)),
-    target=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE)),
-)
-
-__mdmc_prob_preds = torch.rand(NUM_BATCHES, BATCH_SIZE, NUM_CLASSES, EXTRA_DIM)
-__mdmc_prob_preds = __mdmc_prob_preds / __mdmc_prob_preds.sum(dim=2, keepdim=True)
-
-_input_multidim_multiclass_prob = Input(
-    preds=__mdmc_prob_preds,
-    sens=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE, EXTRA_DIM)),
-    target=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE, EXTRA_DIM)),
-)
-
-_input_multidim_multiclass = Input(
-    preds=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE, EXTRA_DIM)),
-    sens=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE, EXTRA_DIM)),
-    target=torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE, EXTRA_DIM)),
-)
-
-
-@pytest.mark.parametrize('sens', [0, 1, 2])
-def test_acc(sens: int) -> None:
-    """Test the Accuracy per sens metric runs."""
-    acc = AccuracyPerSens(sens)
-    acc(_input_binary.preds, _input_binary.sens, _input_binary.target)
-    acc.compute()
-    print(getattr(acc, "__name__", acc.__class__.__name__).lower())
